Get dataset

getDataset imports datasets as TreeSummarizedExperiment objects.

getDataset(x, dryrun = TRUE)

Arguments

x
A character vector with the name(s) of the dataset(s). If empty and dryrun = TRUE, it returns a message with the names of the available datasets. If empty and dryrun = FALSE, it returns a list of TreeSummarizedExperiments with all of the datasets.

dryrun
If TRUE, only returns a message and invisibly returns the names of the datasets as a character vector. If FALSE, it returns the TreeSummarizedExperiment datasets indicated in the argument 'x'.

Value

A list of TreeSummarizedExperiments.

Examples

datasets_names <- getDataset()
#> 1 Beghini_2019_16S_smoking
#> 2 HMP_2012_16S_gingival_V13
#> 3 HMP_2012_16S_gingival_V35
#> 4 HMP_2012_16S_gingival_V35_subset
#> 5 HMP_2012_WMS_gingival
#> 6 Stammler_2016_16S_spikein
#> 
#> Use vignette('datasets', package = 'MicrobiomeBenchmarkData') for a detailed description of the datasets.
#> 
#> Use getDataset(dryrun = FALSE) to import all of the datasets.
datasets <- getDataset("HMP_2012_16S_gingival_V35_subset", dryrun = FALSE)
#> Finished HMP_2012_16S_gingival_V35_subset.