Get Data Sets

getDataSets returns a list containing datasets that have been specified by index according to the output of showDatasets.

getDatasets(x)

Arguments

x
A numeric vector containing the indexes of the datasets to be loaded.

Value

A list of TreeSummarizedExperiments.

See also

Examples

library(MicrobiomeBenchmarkData)
datasets <- getDatasets(x = c(1,3))
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
datasets
#> $Calgaro_2020_16S_Stool_TongueDorsum
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> class: TreeSummarizedExperiment 
#> dim: 1298 78 
#> metadata(0):
#> assays(1): counts
#> rownames(1298): OTU_97.37504 OTU_97.34346 ... OTU_97.45365 OTU_97.45307
#> rowData names(6): SUPERKINGDOM PHYLUM ... FAMILY GENUS
#> colnames(78): 700023058 700023227 ... 700114000 700114717
#> colData names(7): RSID VISITNO ... HMP_BODY_SUBSITE SRS_SAMPLE_ID
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> reducedDimNames(0):
#> mainExpName: NULL
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> altExpNames(0):
#> rowLinks: a LinkDataFrame (1298 rows)
#> rowTree: 1 phylo tree(s) (1298 leaves)
#> colLinks: NULL
#> colTree: NULL
#> 
#> $Calgaro_2020_16S_Subgingival_Supragingival
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> class: TreeSummarizedExperiment 
#> dim: 892 76 
#> metadata(0):
#> assays(1): counts
#> rownames(892): OTU_97.31247 OTU_97.44487 ... OTU_97.45365 OTU_97.45307
#> rowData names(6): SUPERKINGDOM PHYLUM ... FAMILY GENUS
#> colnames(76): 700023057 700023179 ... 700114009 700114338
#> colData names(7): RSID VISITNO ... HMP_BODY_SUBSITE SRS_SAMPLE_ID
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> reducedDimNames(0):
#> mainExpName: NULL
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> Found more than one class "phylo" in cache; using the first, from namespace 'phyloseq'
#> Also defined by ‘tidytree’
#> altExpNames(0):
#> rowLinks: a LinkDataFrame (892 rows)
#> rowTree: 1 phylo tree(s) (892 leaves)
#> colLinks: NULL
#> colTree: NULL
#>